requirements.txt
streamlit
openai
tavily-python
langchain
langchain-openai
langgraph
langchain-community
dashscope
faiss-cpu
tiktoken
pypdf
rapidocr-onnxruntime